Adds printer for inheritance and updates inheritance-graph printer
